Lint cleanup
