+++ ui/src/message_popup/scoring_page.jsx
@@ -2,37 +2,19 @@
 import _ from 'lodash';
 import React from 'react';
 import request from 'superagent';
-import SwipeableViews from 'react-swipeable-views';
 
-import * as Colors from 'material-ui/styles/colors';
 import Divider from 'material-ui/Divider';
 import AppBar from 'material-ui/AppBar';
 import Paper from 'material-ui/Paper';
-import TextField from 'material-ui/TextField';
-import RaisedButton from 'material-ui/RaisedButton';
-import FlatButton from 'material-ui/FlatButton';
-import DoneIcon from 'material-ui/svg-icons/action/done';
 import SchoolIcon from 'material-ui/svg-icons/social/school';
-import FeedbackIcon from 'material-ui/svg-icons/action/feedback';
-import {Card, CardActions, CardHeader, CardText} from 'material-ui/Card';
-import {List, ListItem, Subheader} from 'material-ui/List';
-import {
-  Table,
-  TableHeader,
-  TableHeaderColumn,
-  TableBody,
-  TableRow,
-  TableRowColumn
-} from 'material-ui/Table';
+import {List, ListItem} from 'material-ui/List';
 
 import * as Routes from '../routes.js';
-import Question from './question.jsx';
-import ScoringSection from './scoring_section.jsx';
 import ScoringSwipe from './scoring_swipe.jsx';
 import {withLearningObjective} from './transformations.jsx';
 
 function hashCode(s){
-  return s.split("").reduce(function(a,b){a=((a<<5)-a)+b.charCodeAt(0);return a&a},0);              
+  return s.split("").reduce(function(a,b){a=((a<<5)-a)+b.charCodeAt(0);return a&a;},0);              
 }
 
 function questionId(question) {
@@ -203,69 +185,5 @@ export default React.createClass({
         onCancel={this.onQuestionSelected.bind(this, null)}
       />
     );
-  },
-
-        /*
-        <List>
-          {logs.map((log) => {
-            return <ListItem
-              key={log.id}
-              onClick={this.onLogSelected.bind(this, log)}
-              primaryText={`#${log.id}   ${log.json.question.text}`}
-              secondaryTextLines={2}
-              secondaryText={log.json.initialResponseText} />
-          })}
-        </List>
-        <Table style={styles.summaryTable}>
-          <TableBody displayRowCheckbox={false}>
-            {logs.map((log) => {
-              return (
-                <TableRow key={log.id} onClick={this.onLogSelected.bind(this, log)}>>
-                  <TableRowColumn>{log.id}</TableRowColumn>
-                  <TableRowColumn>{log.json.question.text}</TableRowColumn>
-                  <TableRowColumn>{log.json.initialResponseText}</TableRowColumn>
-                </TableRow>
-              );
-            })}
-          </TableBody>
-        </Table>
-        */
-
-  // renderQuestion(log) {
-  //   const {id, initialResponseText, elapsedMs, question} = log.json;
-  //   const secondsRemaining = Math.round((elapsedMs) / 1000);
-  //   return <Question
-  //     key={id}
-  //     question={question}
-  //     shouldShowStudentCard={true}
-  //     initialResponseText={initialResponseText}
-  //     disabled={true}
-  //     tickerText={`${secondsRemaining}s to respond`} />;
-  // },
-
-  // renderScoringSection(selectedLog) {
-  //   const {question, initialResponseText, elapsedMs} = selectedLog.json;
-  //   const {learningObjective} = withLearningObjective(question)
-
-  //   return <ScoringSection
-  //       logId={selectedLog.id}
-  //       initialResponseText={initialResponseText}
-  //       elapsedMs={elapsedMs}
-  //       question={question}
-  //       learningObjective={learningObjective}
-  //       onDoneScoring={this.onDoneScoring} />;
-  // }
-});
-
-const styles = {
-  summaryTitle: {
-    padding: 20
-  },
-  summaryTable: {
-    border: '1px solid black',
-    paddingBottom: 20,
-    paddingTop: 20,
-    overflow: 'scroll-y',
-    width: 400
   }
-};
+});
